+++ R/RcppExports.R
@@ -38,20 +38,22 @@
 #' @param regret_array User specified regret array. If specifiec, the regret will not be calculated by profoc.
 #' @param trace If a progessbar shall be printed. Defaults to TRUE.
 #' @param init_weights Matrix of dimension Kx1 or KxP used as starting weights. Kx1 represents the constant solution with equal weights over all P whereas specifiying a KxP matrix allows different starting weights for each P.
+#' @param lead_time offset for expert forecasts. Defaults to 0, which means that
+#' experts forecast t+1 at t. Setting this to h means experts predictions refer
+#' to t+1+h at time t. The weight updates delay accordingly.
 #' @usage profoc(y, experts, tau, ex_post_smooth = FALSE, ex_post_fs = FALSE,
 #' lambda = -Inf, method = "boa", method_var = "A", forget_regret = 0,
 #' forget_performance = 0, fixed_share = 0, gamma = 1, ndiff = 1, deg = 3,
 #'rel_nseg = 0.5, gradient = TRUE, loss_array = NULL, regret_array = NULL,
-#' trace = TRUE,
-#' init_weights = NULL)
+#' trace = TRUE, init_weights = NULL, lead_time = 0)
 #' @return Profoc can tune various parameters automatically based on
 #' the past loss. For this, lambda, forget, fixed_share, gamma, ndiff,
 #' deg and rel_nseg can be specified as numeric vectors containing
 #' parameters to consider. Profoc will automatically try all possible
 #' combinations of values provide.
 #' @export
-profoc <- function(y, experts, tau = as.numeric( c()), ex_post_smooth = FALSE, ex_post_fs = FALSE, lambda = as.numeric( c()), method = "boa", method_var = "A", forget_regret = as.numeric( c()), forget_performance = 0, fixed_share = as.numeric( c()), gamma = as.numeric( c()), ndiff = as.numeric( c()), deg = as.numeric( c()), rel_nseg = as.numeric( c()), gradient = TRUE, loss_array = as.numeric( c()), regret_array = as.numeric( c()), trace = TRUE, init_weights = NULL) {
-    .Call(`_profoc_profoc`, y, experts, tau, ex_post_smooth, ex_post_fs, lambda, method, method_var, forget_regret, forget_performance, fixed_share, gamma, ndiff, deg, rel_nseg, gradient, loss_array, regret_array, trace, init_weights)
+profoc <- function(y, experts, tau = as.numeric( c()), ex_post_smooth = FALSE, ex_post_fs = FALSE, lambda = as.numeric( c()), method = "boa", method_var = "A", forget_regret = as.numeric( c()), forget_performance = 0, fixed_share = as.numeric( c()), gamma = as.numeric( c()), ndiff = as.numeric( c()), deg = as.numeric( c()), rel_nseg = as.numeric( c()), gradient = TRUE, loss_array = as.numeric( c()), regret_array = as.numeric( c()), trace = TRUE, init_weights = NULL, lead_time = 0L) {
+    .Call(`_profoc_profoc`, y, experts, tau, ex_post_smooth, ex_post_fs, lambda, method, method_var, forget_regret, forget_performance, fixed_share, gamma, ndiff, deg, rel_nseg, gradient, loss_array, regret_array, trace, init_weights, lead_time)
 }
 
 #' Spline Fit - Fit Penalized B-Splines