+++ NEWS.md
@@ -3,7 +3,9 @@ profoc v0.5.X (Release date: 2021-XX-XX)
 
 Changes:
 
-* The `forget` parameter was renamed to `forget_regret` to underline its reference to the regret. Additionally `forget_performance` was added. It defines which share of the past performance will not be considered for selecting the best parameter combination.
+* Add `lead_time` parameter. offset for expert forecasts. Defaults to 0, which means that experts predict t+1 at t. Setting this to h means experts predictions refer to t+1+h at time t. The weight updates delay accordingly.
+* `forget_performance` was added. It defines which share of the past performance will not be considered for selecting the best parameter combination.
+* The `forget` parameter was renamed to `forget_regret` to underline its reference to the regret.
 * The `init_weights` parameter has to be either a Kx1 or Kx99 matrix specifying the experts' starting weights.
 * A new `trace` parameter lets you decide whether a progress bar is printed.
 * `tau` is now optional. It defaults to 1:P/(P+1). A scalar given to tau will be repeated P times. The latter is useful in multivariate settings.